pyproject.toml
[tool.poetry]
name = "wordgame"
version = "0.1.0"
description = ""
authors = ["aleb2000 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
Flask = "^2.0.2"
pyenchant = "^3.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
